+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 80.0 - 2024-11-18
+## 80.0 - 2024-12-17
 ### Added
 - Achievements added in 11.0.7
 - Feature to prioritize which tab is used when getting redirected to an achievement